Moved SparkVersionUtils to SparkUtils.
