+++ connector/src/main/scala/com/vertica/spark/util/spark/SparkUtils.scala
@@ -13,8 +13,21 @@
 
 package com.vertica.spark.util.spark
 
+import com.vertica.spark.util.version.Version
 import org.apache.spark.sql.SparkSession
+
+import scala.util.Try
 
 object SparkUtils {
   def getVersionString: Option[String] = SparkSession.getActiveSession.map(_.version)
+
+  def getVersion(versionStr: Option[String]): Option[Version] = versionStr match {
+    case Some(str) =>
+      val regex = "([0-9]+)\\.([0-9]+)\\.([0-9]+)(.*)".r
+      Try{
+        val regex(major, minor, service, _) = str
+        Some(Version(major.toInt, minor.toInt, service.toInt))
+      }.getOrElse(None)
+    case None => None
+  }
 }
